Enhancement #10 - Add testing for ApiService class
Added the necessary changes for testing of the ApiService.

Changed the httpclient to use DI to allow for mocking. Added the mockhttp package to make mocking easier.

Added teh ApiServiceTest class for the tests and removed the default UnitTest1 class.

Various cleanup to code.

Updated the readme to include more design decisions